+++ linear_system/matrread.h
@@ -1,2 +1,2 @@
-int read_matr_file(float* A, float* b, int n, char* filename);
-int read_matr_formula(float* A, float* b, int n);
+int read_matr_file(double* A, double* b, int n, char* filename);
+int read_matr_formula(double* A, double* b, int n);
