+++ linear_system/utils.h
@@ -1,6 +1,6 @@
-float* getel(float* A, int n, int row, int col);
-void print_solution(float* x, int n, int m);
-void print_error(float* A, int n, float* b, float* x);
-void print_system(float* A, float* b, int n);
-void print_solution_error(float* x, int n);
-int is_zero(float x);
+double* getel(double* A, int n, int row, int col);
+void print_solution(double* x, int n, int m);
+void print_error(double* A, int n, double* b, double* x);
+void print_system(double* A, double* b, int n);
+void print_solution_error(double* x, int n);
+int is_zero(double x);
